+++ packages/api/src/agents/index.ts
@@ -1,5 +1,6 @@
 export * from './avatars';
 export * from './chain';
+export * from './client';
 export * from './context';
 export * from './edges';
 export * from './handlers';
