+++ packages/api/src/types/http.ts
@@ -1,5 +1,6 @@
-import type { Request } from 'express';
 import type { IUser, AppConfig } from '@librechat/data-schemas';
+import type { TEndpointOption } from 'librechat-data-provider';
+import type { Request } from 'express';
 
 /**
  * LibreChat-specific request body type that extends Express Request body
@@ -11,8 +12,10 @@ export type RequestBody = {
   conversationId?: string;
   parentMessageId?: string;
   endpoint?: string;
+  endpointType?: string;
   model?: string;
   key?: string;
+  endpointOption?: Partial<TEndpointOption>;
 };
 
 export type ServerRequest = Request<unknown, unknown, RequestBody> & {
